pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roland"
version = "0.1.0"
description = "Voice-controlled copilot for Star Citizen with JARVIS-style AI assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Chris"}
]
keywords = ["star-citizen", "voice-assistant", "gaming", "ai", "tts", "stt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Simulation",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Speech-to-Text
    "RealtimeSTT>=0.1.15",
    "faster-whisper>=1.0.0",
    "openwakeword>=0.6.0",

    # Text-to-Speech
    "TTS>=0.22.0",

    # LLM
    "ollama>=0.3.0",
    "httpx>=0.27.0",

    # Keyboard
    "pynput>=1.7.6",

    # System Tray
    "pystray>=0.19.5",
    "Pillow>=10.0.0",

    # Audio
    "PyAudio>=0.2.14",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",

    # Storage
    "aiosqlite>=0.20.0",

    # Configuration
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
roland = "roland.main:main"

[project.urls]
Homepage = "https://github.com/chris/roland"
Repository = "https://github.com/chris/roland"
Issues = "https://github.com/chris/roland/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["roland*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=roland --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
